Do not reset selection on :view command.

Thanks to Alexandre Viau.

Fix selection created prior to running av in normal mode showing up
after switching to a regular visual mode via v key and moving cursor
over it.

Fix :keepsel not preserving statusbar messages.

Fix mborder top line bit flickering on startup.

Don't draw inactive cursor mark at load_stage 1 which makes for smoother
initial drawing.

Update README of Vim plugin with :PeditVifm.

Plugins
-------

* Schedule redraw after VifmView:cd() to fix tabline.

+++ src/utils/utils.c
@@ -318,7 +318,7 @@ split_size_double(double size, unsigned long long *ifraction,
 	*fraction_width = (cfg.sizefmt.precision == 0) ? 1 : cfg.sizefmt.precision;
 
 	ten_power = pow(10, *fraction_width + 1);
-	while(ten_power > ULLONG_MAX)
+	while(ten_power > (double)ULLONG_MAX)
 	{
 		ten_power /= 10;
 		--*fraction_width;
